microservice/requirements.txt
fastapi[all]
uvicorn[standard]
python-dotenv==1.0.1
pydantic-settings==2.6.1
pandas==2.2.3
pypdf2==3.0.1
python-docx==1.1.2
pytesseract==0.3.13
pillow==11.0.0
openpyxl==3.1.5
python_multipart==0.0.17
orjson==3.10.12
loguru
SQLAlchemy==2.0.36
asyncpg==0.30.0